+++ docs/source/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `moldrug.utils.Individual` now is a hashable object.
+- `moldrug.utils.GA.SawIndividuals` now is a `set` instead of a `list`
 
 ## [1.0.2] - 2022.08.08